cloud/proxy/requirements.txt
aiortc==1.14.0
av==16.0.1
grpcio==1.76.0
grpcio-tools==1.76.0
# numpy==2.3.5
opencv-python-headless==4.12.0.88
pandas==2.3.3
protobuf==6.33.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
redis==7.1.0
